fix(monk): 무도가 전직 기술을 원작 셋(이형환위·단각·양의신권)으로 바꾼다
ClassChooser 의 무도가 갈래가 영어 기술 여섯 줄(Wolf Fang Fist 등, Claw Fist 중복
포함)을 주고 있어 원작 기술이 하나도 들어오지 않았다. 이형환위·단각·양의신권을
대신 준다. 양의신권은 ScriptName·Type 이 기본공격("Assail")과 같아, 생성 때
이미 받은 영어 Assail 을 먼저 지워야 평타가 두 번 나가지 않는다 — 그래서
무도가로 고르는 순간 기본공격 자리가 양의신권으로 바뀐다.

쿨타임을 0.25초 단위로 받으려고 SkillTemplate.Cooldown 을 int 에서 double 로
바꿨다. AddSeconds(double) 를 쓰는 자리는 그대로 통하고, ServerFormat3F(정수
초만 보내는 화면 표시 패킷)로 보내는 자리 하나만 (int) 로 깎아 보냈다 — 실제
쿨타임은 서버가 double 로 그대로 지킨다.

이형환위 쿨타임 7→0.25초, 단각 4→2초. 양의신권은 기본공격이라 0 그대로 둔다.

+++ src/Hades.Server.Base/Network/Game/GameClient.cs
@@ -1196,7 +1196,7 @@ namespace Darkages.Network.Game
 
             Send(new ServerFormat3F(1,
                 skill.Slot,
-                skill.Template.Cooldown));
+                (int)skill.Template.Cooldown));
         }
 
         public void TrainSpell(Spell spell)
